+++ roguelike/Vec2d.cpp
@@ -3,7 +3,7 @@
 Vec2d::Vec2d() : x(), y() {
 }
 
-Vec2d::Vec2d(uint32_t x, uint32_t y) : x(x), y(y) {
+Vec2d::Vec2d(size_t x, size_t y) : x(x), y(y) {
 }
 
 Vec2d::Vec2d(const Vec2d& other) : x(other.x), y(other.y) {
@@ -29,12 +29,12 @@ Vec2d& Vec2d::operator-=(const Vec2d& rhs) {
 	return *this;
 }
 
-uint32_t Vec2d::length() {
+size_t Vec2d::length() {
 	return this->x + this->y;
 }
 
-uint32_t Vec2d::distance(const Vec2d& lhs, const Vec2d& rhs) {
-	uint32_t dist = 0;
+size_t Vec2d::distance(const Vec2d& lhs, const Vec2d& rhs) {
+	size_t dist = 0;
 	if (lhs.x > rhs.x) {
 		dist += lhs.x - rhs.x;
 	} else {
